backend/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
playwright==1.52.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
